changes to input to allow fragment-style geometry specification fixed some gcc warnings with "(int) ceil"
git-svn-id: https://sirius.chem.vt.edu/svn/psi3/trunk@3818 29f243b5-bcef-41bb-bbe7-8e20262f3206

+++ src/lib/libdpd/buf4_axpy.cc
@@ -47,7 +47,7 @@ int dpd_buf4_axpy(dpdbuf4 *BufX, dpdbuf4 *BufY, double alpha)
 
       if(!rows_per_bucket) dpd_error("buf4_axpy: Not enough memory for one row!", stderr);
 
-      nbuckets = ceil(((double) BufX->params->rowtot[h])/((double) rows_per_bucket));
+      nbuckets = (int) ceil(((double) BufX->params->rowtot[h])/((double) rows_per_bucket));
 
       rows_left = BufX->params->rowtot[h] % rows_per_bucket;
